+++ docs/changelog.rst
@@ -14,10 +14,12 @@ Added
 - Catalog specifically for HydroMT-Wflow.
 - new function **setup_1dmodel_connection** to connect wflow to 1D river model (eg Delft3D FM 1D, HEC-RAS, etc.) `PR #210 <https://github.com/Deltares/hydromt_wflow/pull/210>`_
 - Setup method for the **KsatHorFrac** parameter. PR #249
+- workflow to compute brooks corey c for the wflow layers based on soilgrids data, soilgrids_brooks_corey. PR #242
 - Workflow function **setup_ksathorfarc** to up-downscale existing ksathorfrac maps. PR #249
 
 Changed
 -------
+- **setup_soilmaps**: the user can now supply variable sbm soil layer thicknesses. The Brooks Corey coefficient is then computed as weighted average over the sbm layers. PR #242
 - **setup_outlets**: the IDs of wflow_subcatch are used to define the outlets IDs rather than [1:n]. PR #247
 
 Fixed